Polish stops popup view according to design
