+++ big-red-shuttle/Assets.xcassets/RedPinIcon.imageset/RedPinIcon.png
