+++ big-red-shuttle/Assets.xcassets/RedPinIcon.imageset/RedPinIcon@0.5x.png
