+++ big-red-shuttle/Assets.xcassets/RedPinIcon.imageset/RedPinIcon@1.5x.png
