+++ big-red-shuttle/Network.swift
@@ -2,10 +2,6 @@ import Alamofire
 import SwiftyJSON
 import CoreLocation
 
-struct Constants {
-    static let brsStackBaseURL = "https://raw.githubusercontent.com/cuappdev/big-red-shuttle-stack/master/"
-}
-
 func system() -> System {
     return System.shared
 }
@@ -140,171 +136,4 @@ class API {
             topViewController.present(alertController, animated: true, completion: nil)
         }
     }
-}
-
-// Get BRS stops with name, location, dates, and times
-public func getStops() ->  [Stop] {
-    let stopJsonURLString = "\(Constants.brsStackBaseURL)brs-stops.json"
-    var stops = [Stop]()
-    if let url = URL(string: stopJsonURLString) {
-        if let data = try? Data(contentsOf: url) {
-            let json = JSON(data: data)
-            for stop in json["stops"].arrayValue {
-                var dayArray = [Days]()
-                let name = stop["name"].stringValue
-                let lat = stop["lat"].floatValue
-                let long = stop["long"].floatValue
-                let days = stop["days"].arrayObject!
-                let timeStrings = stop["times"].arrayObject! as! [String]
-                
-                var times:[Time] = []
-                for day in days {
-                    let dayObject = Days(rawValue: day as! String)!
-                    dayArray.append(Days(rawValue: day as! String)!)
-                    for tStr in timeStrings {
-                        times.append(Time(time: tStr, technicallyNightBefore: dayObject.number))
-                    }
-                }
-                stops.append(Stop(name: name, lat: lat, long: long, days: dayArray, times: times))
-                
-            }
-        }
-    }
-    return stops
-}
-
-// Get BRS emergency contacts with service and name
-public func getEmergencyContacts() -> [EmergencyContact] {
-    let emergencyContactsURLString = "\(Constants.brsStackBaseURL)brs-emergency-contacts.json"
-    var emergencyContacts = [EmergencyContact]()
-    
-    if let url = URL(string: emergencyContactsURLString) {
-        if let data = try? Data(contentsOf: url) {
-            let json = JSON(data: data)
-            for contact in json["contacts"].arrayValue {
-                let service = contact["service"].stringValue
-                let number = contact["number"].stringValue
-
-                emergencyContacts.append(EmergencyContact(service: service, number: number))
-            }
-        }
-    }
-    
-    return emergencyContacts
-}
-
-// Get descriptions for BRS About View
-public func getAboutSections() -> [AboutSection] {
-    let aboutSectionsURLString = "\(Constants.brsStackBaseURL)brs-about-sections.json"
-    var aboutSections = [AboutSection]()
-    
-    if let url = URL(string: aboutSectionsURLString) {
-        if let data = try? Data(contentsOf: url) {
-            let json = JSON(data: data)
-
-            for section in json["sections"].arrayValue {
-                let title = section["title"].stringValue
-                let summary = section["summary"].stringValue
-                let link = section["link"].stringValue
-                
-                aboutSections.append(AboutSection(title: title, summary: summary, link: link))
-            }
-        }
-    }
-    
-    return aboutSections
-}
-
-public func getUniqueStops() -> [Stop] {
-    var names:[String] = []
-    var uniqueStops:[Stop] = []
-    
-    for stop in getStops() {
-        if !names.contains(stop.name) {
-            uniqueStops.append(stop)
-            names.append(stop.name)
-        }
-    }
-    
-    return uniqueStops
-}
-
-class Polyline: NSObject {
-    var overviewPolyline = ""
-    
-    enum Status: String {
-        case OK = "OK"
-        case ZERO_RESULTS = "ZERO_RESULTS"
-        case OVER_QUERY_LIMIT = "OVER_QUERY_LIMIT"
-        case REQUEST_DENIED = "REQUEST_DENIED"
-        case INVALID_REQUEST = "INVALID_REQUEST"
-        case UNKNOWN_ERROR = "UNKNOWN_ERROR"
-    }
-    
-    func getPolyline(waypoints:[Stop], origin:Stop, end:Stop) {
-        let baseURLDirections = "https://maps.googleapis.com/maps/api/directions/json?"
-        let origin = "\(origin.lat),\(origin.long)"
-        let destination = "\(end.lat),\(end.long)"
-        
-        var directionsURLString = baseURLDirections + "origin=" + origin + "&destination=" + destination
-        
-        if waypoints.count > 0 {
-            directionsURLString += "&waypoints=optimize:true"
-            for i in 0..<waypoints.count {
-                let coords = "\(waypoints[i].lat),\(waypoints[i].long)"
-                directionsURLString +=  "|" + coords
-            }
-        }
-        
-        directionsURLString = directionsURLString.addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed)!
-        print(directionsURLString)
-        if let directionsURL = URL(string: directionsURLString) {
-            if let data = try? Data(contentsOf: directionsURL) {
-                let json = JSON(data: data)
-                let status = Status(rawValue: json["status"].stringValue)!
-                
-                switch status {
-                    case .OK:
-                        overviewPolyline = json["routes"][0]["overview_polyline"]["points"].stringValue
-                    case .ZERO_RESULTS:
-                        print("zero results, can't draw bus routes")
-                    case .OVER_QUERY_LIMIT:
-                        print("over query limit, can't draw bus routes")
-                    case .REQUEST_DENIED:
-                        print("request was denied, can't draw bus routes")
-                    case .INVALID_REQUEST:
-                        print("invalid request, can't draw bus routes")
-                    case .UNKNOWN_ERROR:
-                        print("unknown, can't draw bus routes")
-                }
-            }
-        }
-    }
-}
-
-
-/* Creates time from JSON string
- * Time string must have format:
- * H:mm a
- */
-public func getTime(time: String) -> (Int, Int) {
-    var hourArr = time.components(separatedBy: ":") //["h","mm a"]
-    let minArr = hourArr[1].components(separatedBy: " ") //["mm","a"]
-    let timeArr = [hourArr[0]] + minArr //["h","mm","a"]
-    
-    let minute = Int(timeArr[1])!
-    
-    let hour12 = Int(hourArr[0])!
-    
-    var hour = 0
-    let aa = timeArr[2].lowercased()
-    if(aa=="am" && hour12 == 12){ //12:xx am = 0:xx
-        hour = 0
-    }else if(aa=="pm" && hour12 != 12){ //12:xx pm = 12:xx
-        hour = hour12 + 12 //h:xx pm = (h+12):xx
-    }else{
-        hour = hour12
-    }
-
-    return (hour, minute)
 }
