+++ big-red-shuttle/UIColor+Shared.swift
@@ -27,4 +27,5 @@ extension UIColor {
     @nonobjc static let searchbordergray = UIColor(red: 25/255, green: 25/255, blue: 25/255, alpha: 0.5)
     @nonobjc static let searchbottombordergray = UIColor(red: 224/255, green: 224/255, blue: 224/255, alpha: 0.5)
     @nonobjc static let navtitleblack = UIColor(red: 54/255, green: 53/255, blue: 67/255, alpha: 1.0)
+    @nonobjc static let directionsgrey = UIColor(red: 179/255, green: 183/255, blue: 192/255, alpha: 1.0)
 }
